DublinBikes Prediction
Added the DublinBikes_Model directory, which contains 'dublinbikes_24hr_prediction.py'. The 'dublinbikes_prediction()' function contains notes on how to use it.

Input(X, station_ID)
X = a list of the 20 most recent observations
station_ID = the stationID of the station associated with the observations

Output Y
Y = a list containing 24*12 predictions, for each 5-minute interval for the next 24 hours.